Dealing with Vectors
Created program that unwraps a 2D vector, converts the 1D vector to an
array and then reassigns a value in the array.  The change is maintained
when converted back into a 2D vector.

Unwrapping is necessary b/c a 2D vector is a "vector of vectors" which
means that it can be represented as a pointer to a set of pointers to a
set of values (blech this is getting confusing).

Wrap and Unwrap functions will likely be tested to improve performance,
especially when dealing with very large vectors.