Additional testing
Successfully ran testVectorPtr.cpp on large array.
Wrote tests for examining how CUDA treated objects (testCudaObj*).
These tests confirmed that CUDA can receive and work with objects and
arrays with objects.  This is extremely useful as the ability to use
objects makes is significantly easier to manipulate data both in and out
of CUDA.  However, it should be noted that the allocation and copying of
objects between the host and the device is slightly confusing.

In addition, testCudaMultiCpy.cpp was written to test how CUDA treated
memory calls of the same size but different frequency.  In other words,
this bit of code contains 3 arrays of size 10000 and a single array of
size 10000*3.  It was noted that the single array is almost 1000 times
faster when attempting to allocate and copy to host.  This is likely due
to the overhead associated with cudaMemcpy and cudaMalloc.

As a result, the modified GPU code will now try to group as much as
possible under a few rather large (but well organized) objects.