Additional Test within Multi Copy
Added a second test to testCudaMultiCpy.cpp.  This test is represented
by "TIME 3: " and attempts to see if passing a large object followed by
returning a smaller object is more efficient then passing a large object
and returning it.  The current results indicated that yes, passing a
large object and returning a smaller object is faster.  This occurs
despite the need to run additional cudaMalloc commands.